+++ src/elements/GeoCalendar/GeoCalendarGrid/GeoCalendarDayGridDayUnit.vue
@@ -1,6 +1,7 @@
 <template>
   <!-- mousedown event is used because it is fired before blur event on GeoInput -->
   <!-- blur event won't be fired but that's fine because we want this handler to prevail over the blur one -->
+  <!-- https://forum.vuejs.org/t/blur-before-click-only-on-safari/21598/7 -->
   <button
     :class="{
       'geo-calendar-grid__day-unit': true,
@@ -13,7 +14,8 @@
       'geo-calendar-grid__date-picker-unit--to-date': isEqual(day, selectedToDay),
       'geo-calendar-grid__date-picker-unit--within-range': isDayWithinRanges
     }"
-    @mousedown="selectDay($event)"
+    @mousedown.prevent="selectDay($event)"
+    @mouseover="emitDayUnitMouseover($event)"
   >
     <div class="geo-calendar-grid__date-picker-unit__placeholder">
       {{ dayNumber }}
@@ -114,6 +116,16 @@ export default {
        * @type {Date}
        */
       this.$emit('select-day-unit', this.day)
+    },
+
+    emitDayUnitMouseover () {
+      /**
+       * User hovers on a potential selected date
+       *
+       * @event day-unit-mouseover
+       * @type {Date}
+       */
+      this.$emit('day-unit-mouseover', this.day)
     }
   }
 }