+++ src/elements/GeoCalendar/GeoCalendarGrid/GeoCalendarMonthGrid.vue
@@ -14,6 +14,7 @@
         :selected-to-day="selectedToDay"
         @select-month="selectMonth"
         @select-quarter="selectQuarter($event)"
+        @month-unit-mouseover="emitMonthUnitMouseover($event)"
       />
     </div>
   </div>
@@ -74,6 +75,16 @@ export default {
        * @type {Number}
        */
       this.$emit('select-quarter', monthIndex)
+    },
+
+    emitMonthUnitMouseover (monthIndex) {
+      /**
+       * User hovers on a potential selected month
+       *
+       * @event month-unit-mouseover
+       * @type {Number}
+       */
+      this.$emit('month-unit-mouseover', monthIndex)
     }
   }
 }