+++ src/elements/GeoCalendar/GeoCalendarGrid/GeoCalendarMonthGridMonthUnit.vue
@@ -1,6 +1,7 @@
 <template>
   <!-- mousedown event is used because it is fired before blur event on GeoInput -->
   <!-- blur event won't be fired but that's fine because we want this handler to prevail over the blur one -->
+  <!-- https://forum.vuejs.org/t/blur-before-click-only-on-safari/21598/7 -->
   <button
     :class="{
       'geo-calendar-grid__month-unit': true,
@@ -11,7 +12,8 @@
       'geo-calendar-grid__date-picker-unit--from-date': isDayWithinFromMonth,
       'geo-calendar-grid__date-picker-unit--to-date': isDayWithinToMonth
     }"
-    @mousedown="selectMonth($event)"
+    @mousedown.prevent="selectMonth($event)"
+    @mouseover="emitMonthUnitMouseover($event)"
   >
     <div class="geo-calendar-grid__date-picker-unit__placeholder">
       {{ monthName }}
@@ -114,6 +116,16 @@ export default {
        * @type {Number}
        */
       this.$emit('select-month-unit', this.monthIndex)
+    },
+
+    emitMonthUnitMouseover () {
+      /**
+       * User hovers on a potential selected month
+       *
+       * @event month-unit-mouseover
+       * @type {Number}
+       */
+      this.$emit('month-unit-mouseover', this.monthIndex)
     }
   }
 }