+++ src/elements/GeoInput/GeoInput.vue
@@ -24,6 +24,16 @@
         v-on="listeners"
         @input="onInput($event)"
       >
+      <!-- mousedown event is used because it is fired before blur event on GeoInput -->
+      <!-- blur event won't be fired but that's fine because we want this handler to prevail over the blur one -->
+      <!-- https://forum.vuejs.org/t/blur-before-click-only-on-safari/21598/7 -->
+      <font-awesome-icon
+        v-if="!disabled && !!value"
+        :icon="deleteInputValueIcon"
+        fixed-with
+        class="geo-input__icon geo-input__icon--trailing geo-input__icon--delete"
+        @mousedown.prevent="deleteValue"
+      />
       <font-awesome-icon
         v-if="disabled"
         :icon="disabledIcon"
@@ -104,6 +114,19 @@ export default {
     leadingAccessoryIcon: {
       type: Array,
       required: false
+    },
+
+    /**
+     * Font Awesome 5 icon to be displayed.
+     *
+     * See [vue-fontawesome](https://www.npmjs.com/package/@fortawesome/vue-fontawesome#explicit-prefix-note-the-vue-bind-shorthand-because-this-uses-an-array)
+     * for more info about this.
+     */
+    deleteInputValueIcon: {
+      type: Array,
+      default () {
+        return ['fas', 'times-circle']
+      }
     }
   },
   computed: {
@@ -128,6 +151,15 @@ export default {
        * @type {KeyboardEvent}
        */
       this.$emit('input', $event.target.value)
+    },
+
+    deleteValue () {
+      /**
+       * User typed on the input box.
+       *
+       * @event delete-value
+       */
+      this.$emit('delete-value')
     }
   }
 }